Move testing-related Node.js modules into the testing subdirectory
Part of a larger effort to organize the code in the `src` directory.
