+++ src/index.js
@@ -8,9 +8,9 @@ exports.syncFunctionLoader = require('./sync-function-loader.js');
 /**
  * The test-helper module. Provides a number of conveniences to test the behaviour of document definitions.
  */
-exports.testHelper = require('./test-helper.js');
+exports.testHelper = require('./testing/test-helper.js');
 
 /**
  * The validation-error-formatter module. Formats document validation error messages for use in document definition tests.
  */
-exports.validationErrorFormatter = require('./validation-error-formatter.js');
+exports.validationErrorFormatter = require('./testing/validation-error-formatter.js');
