+++ src/testing/validation-error-formatter.js
