+++ src/testing/validation-error-formatter.spec.js
